refactor: clean naming/types and fix spelling in source
