+++ src/client/api/NodeApi.ts
@@ -3,7 +3,7 @@ import { User } from 'blaise-api-node-client/lib/cjs/interfaces/user';
 import { Survey } from '../../common/interfaces/surveyInterface';
 import { SupervisorInformation } from '../types/supervisorInterface';
 import { EditorInformation } from '../types/editorInterface';
-import mapEditorInformation from '../utils/editorInformaitionMapper';
+import mapEditorInformation from '../utils/editorInformationMapper';
 import mapSupervisorInformation from '../utils/supervisorInformationMapper';
 import { CaseSummaryDetails } from '../../common/interfaces/caseInterface';
 import mapCasesNotAllocated from '../utils/caseAllocationMapper';
