+++ src/client/utils/editorInformationMapper.ts
