+++ src/server/questionnaireMapper.ts
@@ -16,12 +16,12 @@ function mapFieldPeriod(fieldPeriodDate: string | undefined): string {
 }
 
 export default function mapQuestionnaireDetails(questionnaire: Questionnaire): QuestionnaireDetails {
-  const questionaireDetails: QuestionnaireDetails = {
+  const questionnaireDetails: QuestionnaireDetails = {
     questionnaireName: questionnaire.name,
     numberOfCases: questionnaire.dataRecordCount ?? 0,
     fieldPeriod: mapFieldPeriod(questionnaire.fieldPeriod),
     surveyTla: questionnaire.surveyTla ?? 'N/A',
   };
 
-  return questionaireDetails;
+  return questionnaireDetails;
 }
